requirements.txt
asgiref==3.8.1
Django==4.2.27
python-dotenv==1.0.1
sqlparse==0.5.5

# PostgreSQL adapter
psycopg2-binary==2.9.9

# Base de datos URL parser
dj-database-url==2.2.0

# Para producción
gunicorn==21.2.0
whitenoise==6.6.0
requests==2.31.0
python-dateutil==2.8.2
